3d-model: track how vertices are arranged separately from their definitions
Most models use the same vertex in more than one (adjacent) polygon, so
tracking the layout of the vertices separately from the vertex
definitions is a good idea.  It will prevent the need from having
redundant entries in the vertex buffer.

This commit switches the drawing code over to using an element array to
that end.